Use 32bit output.

+++ gfx/lighting.cpp
@@ -19,7 +19,7 @@ Lighting::Lighting(const Size<int>& renderSize, const gl::Texture& texDepth) :
 {
     // Texture and FBO
     auto fboSize = {renderSize.w, renderSize.h};
-    tex.bind().alloc(fboSize, GL_RGB16F,  GL_RGB, GL_FLOAT);
+    tex.bind().alloc(fboSize, GL_RGB32F,  GL_RGB, GL_FLOAT);
     fbo.bind()
        .attach(texDepth, gl::Fbo::Attachment::Depth)
        .attach(tex, gl::Fbo::Attachment::Color)
